without of additional task filter contacts for favorite

+++ controllers/contactsControllers.js
@@ -4,7 +4,16 @@ import HttpError from "../helpers/HttpError.js";
 
 export const getAllContacts = async (req, res, next) => {
   try {
-    const result = await contactsService.listContacts();
+    const { _id: owner } = req.user;
+    const { page = 1, limit = 20 } = req.query;
+    const skip = (page - 1) * limit;
+    const result = await contactsService.listContacts(
+      { owner },
+      {
+        skip,
+        limit,
+      }
+    );
     res.json(result);
   } catch (error) {
     next(error);
@@ -13,8 +22,9 @@ export const getAllContacts = async (req, res, next) => {
 
 export const getOneContact = async (req, res, next) => {
   try {
+    const { _id: owner } = req.user;
     const { id } = req.params;
-    const result = await contactsService.getContactById(id);
+    const result = await contactsService.getContactById({ _id: id, owner });
     if (!result) {
       throw HttpError(404);
     }
@@ -26,8 +36,9 @@ export const getOneContact = async (req, res, next) => {
 
 export const deleteContact = async (req, res, next) => {
   try {
+    const { _id: owner } = req.user;
     const { id } = req.params;
-    const result = await contactsService.removeContact(id);
+    const result = await contactsService.removeContact({ _id: id, owner });
     if (!result) {
       throw HttpError(404);
     }
@@ -39,7 +50,8 @@ export const deleteContact = async (req, res, next) => {
 
 export const createContact = async (req, res, next) => {
   try {
-    const result = await contactsService.addContact(req.body);
+    const { _id: owner } = req.user;
+    const result = await contactsService.addContact({ ...req.body, owner });
     res.json(result);
   } catch (error) {
     next(error);
@@ -48,8 +60,12 @@ export const createContact = async (req, res, next) => {
 
 export const updateContact = async (req, res, next) => {
   try {
+    const { _id: owner } = req.user;
     const { id } = req.params;
-    const result = await contactsService.updatContactById(id, req.body);
+    const result = await contactsService.updatContactById(
+      { _id: id, owner },
+      req.body
+    );
     if (!result) {
       throw HttpError(404);
     }
@@ -65,7 +81,11 @@ export const updateContact = async (req, res, next) => {
 export const updateStatusContact = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const result = await contactsService.updatContactByFavorite(id, req.body);
+    const { _id: owner } = req.user;
+    const result = await contactsService.updatContactByFavorite(
+      { _id: id, owner },
+      req.body
+    );
     if (!result) {
       throw HttpError(404);
     }
